requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
anthropic
